Use elfutils instead of libbacktrace for symbolication
Remove the internal libbacktrace copy and instead rely on an
externally provided elfutils and its libdwfl for symbolication.

The code builds on the lessons learned while developing and optimizing
symbolication for perfparser and hotspot. As such, the new code here
directly comes with a bunch of caches and complexity. But overall,
I only see a lot of benefits from this approach:

- we don't need to manually update libbacktrace
- some symbols are now much better symbolicated, most notably for
  lambda expressions and things in anonymous namespaces
- we can leverage this as a good basis for offline symbolication
  with a proper sysroot setup, similar to what we can do in perfparser
- the new symbolication code is significantly faster, in some of my
  local tests heaptrack_interpret now only takes 2.1s instead of 3.3s
  which is quite a noticeable improvement imo